requirements.txt
Flask==2.2.5
Flask-Cors==4.0.0
python-dateutil==2.8.2
python-dotenv==0.21.1
requests==2.31.0
six==1.16.0

